deploy/code/requirements.txt
ultralytics
Pillow